+++ developer-extension/src/panel/hooks/useEvents/eventFilters.spec.ts
@@ -1,7 +1,70 @@
+import type { RumEvent } from '../../../../../packages/rum-core/src/rumEvent.types'
+import type { LogsEvent } from '../../../../../packages/logs/src/logsEvent.types'
 import { isSafari } from '../../../../../packages/core/src/tools/utils/browserDetection'
-import { parseQuery, matchWithWildcard } from './eventFilters'
+import { parseQuery, matchWithWildcard, filterFacets } from './eventFilters'
+import type { FacetValuesFilter } from './eventFilters'
+import { FacetRegistry } from './facetRegistry'
+const RUM_ERROR_EVENT = { type: 'error' } as RumEvent
+const RUM_ACTION_EVENT = { type: 'action' } as RumEvent
+const LOGS_EVENT = { status: 'info', origin: 'logger' } as LogsEvent
+const RUM_XHR_RESOURCE_EVENT = { type: 'resource', resource: { type: 'xhr' } } as RumEvent
 
 if (!isSafari()) {
+  describe('filterFacets', () => {
+    const facetRegistry = new FacetRegistry()
+    facetRegistry.addEvent(RUM_ACTION_EVENT)
+    facetRegistry.addEvent(RUM_ERROR_EVENT)
+    facetRegistry.addEvent(RUM_ERROR_EVENT)
+    facetRegistry.addEvent(RUM_XHR_RESOURCE_EVENT)
+    facetRegistry.addEvent(LOGS_EVENT)
+
+    it('should exclude selected facets when in exclusion mode', () => {
+      expect(
+        filterFacets(
+          [RUM_ACTION_EVENT, RUM_ERROR_EVENT, RUM_ERROR_EVENT],
+          { type: 'exclude', facetValues: { type: ['error'] } } as FacetValuesFilter,
+          facetRegistry
+        )
+      ).toEqual([RUM_ACTION_EVENT])
+    })
+    it('should exclude unselected facets when in inclusion mode', () => {
+      expect(
+        filterFacets(
+          [RUM_ACTION_EVENT, RUM_ERROR_EVENT, RUM_ERROR_EVENT],
+          { type: 'include', facetValues: { type: ['error'] } } as FacetValuesFilter,
+          facetRegistry
+        )
+      ).toEqual([RUM_ERROR_EVENT, RUM_ERROR_EVENT])
+    })
+    it('should include selected facets at different levels in inclusion mode', () => {
+      expect(
+        filterFacets(
+          [RUM_ACTION_EVENT, RUM_ERROR_EVENT, RUM_ERROR_EVENT, RUM_XHR_RESOURCE_EVENT],
+          {
+            type: 'include',
+            facetValues: {
+              type: ['action'],
+              'resource.type': ['xhr'],
+            },
+          } as FacetValuesFilter,
+          facetRegistry
+        )
+      ).toEqual([RUM_ACTION_EVENT, RUM_XHR_RESOURCE_EVENT])
+    })
+    it('should exclude facets at different levels in exclusion mode', () => {
+      const facetValuesFilter: FacetValuesFilter = {
+        type: 'exclude',
+        facetValues: {
+          type: ['action', 'resource', 'view'],
+          $eventSource: ['logs'],
+        },
+      }
+      expect(filterFacets([RUM_ACTION_EVENT, RUM_ERROR_EVENT, LOGS_EVENT], facetValuesFilter, facetRegistry)).toEqual([
+        RUM_ERROR_EVENT,
+      ])
+    })
+  })
+
   describe('parseQuery', () => {
     it('return a simple field', () => {
       expect(parseQuery('foo:bar')).toEqual([['foo', 'bar']])