+++ packages/core/src/index.ts
@@ -103,7 +103,6 @@ export {
   deleteCookie,
   resetInitCookies,
 } from './browser/cookie'
-export { generateAnonymousId } from './domain/user'
 export type { CookieStore, WeakRef, WeakRefConstructor } from './browser/browser.types'
 export type { XhrCompleteContext, XhrStartContext } from './browser/xhrObservable'
 export { initXhrObservable } from './browser/xhrObservable'
@@ -143,7 +142,7 @@ export {
   getSyntheticsResultId,
 } from './domain/synthetics/syntheticsWorkerValues'
 export type { User } from './domain/user'
-export { checkUser, sanitizeUser } from './domain/user'
+export { checkUser } from './domain/user'
 export * from './domain/resourceUtils'
 export * from './tools/utils/polyfills'
 export * from './tools/utils/numberUtils'