+++ packages/rum-core/src/boot/startRum.ts
@@ -34,6 +34,7 @@ import type { RumSessionManager } from '../domain/rumSessionManager'
 import { startRumSessionManager, startRumSessionManagerStub } from '../domain/rumSessionManager'
 import { startRumBatch } from '../transport/startRumBatch'
 import { startRumEventBridge } from '../transport/startRumEventBridge'
+import type { UrlContexts } from '../domain/contexts/urlContexts'
 import { startUrlContexts } from '../domain/contexts/urlContexts'
 import { createLocationChangeObservable } from '../browser/locationChangeObservable'
 import type { RumConfiguration } from '../domain/configuration'
@@ -133,7 +134,7 @@ export function startRum(
   const locationChangeObservable = createLocationChangeObservable(configuration, location)
   const pageStateHistory = startPageStateHistory(configuration)
   const viewHistory = startViewHistory(lifeCycle)
-  const urlContexts = startUrlContexts(lifeCycle, hooks, locationChangeObservable, location, viewHistory)
+  const urlContexts = startUrlContexts(lifeCycle, hooks, locationChangeObservable, location)
 
   const { observable: windowOpenObservable, stop: stopWindowOpen } = createWindowOpenObservable()
   cleanupTasks.push(stopWindowOpen)
@@ -151,6 +152,7 @@ export function startRum(
     domMutationObservable,
     featureFlagContexts,
     windowOpenObservable,
+    urlContexts,
     viewHistory,
     getCommonContext,
     reportError
@@ -250,12 +252,14 @@ export function startRumEventCollection(
   domMutationObservable: Observable<void>,
   featureFlagContexts: FeatureFlagContexts,
   windowOpenObservable: Observable<void>,
+  urlContexts: UrlContexts,
   viewHistory: ViewHistory,
   getCommonContext: () => CommonContext,
   reportError: (error: RawError) => void
 ) {
   const actionCollection = startActionCollection(
     lifeCycle,
+    hooks,
     domMutationObservable,
     windowOpenObservable,
     configuration,
@@ -272,7 +276,7 @@ export function startRumEventCollection(
     hooks,
     sessionManager,
     viewHistory,
-    actionCollection.actionContexts,
+    urlContexts,
     displayContext,
     featureFlagContexts,
     getCommonContext,