+++ packages/rum-core/src/rumEvent.types.ts
@@ -516,7 +516,7 @@ export type RumLongTaskEvent = CommonProperties &
         /**
          * The script resource name where available (or empty if not found)
          */
-        readonly source_url?: string
+        source_url?: string
         /**
          * The script function name where available (or empty if not found)
          */
@@ -528,7 +528,7 @@ export type RumLongTaskEvent = CommonProperties &
         /**
          * Information about the invoker of the script
          */
-        readonly invoker?: string
+        invoker?: string
         /**
          * Type of the invoker of the script
          */
@@ -1382,7 +1382,7 @@ export interface CommonProperties {
     /**
      * Cellular connection type reflecting the measured network performance
      */
-    readonly effective_type?: 'slow_2g' | '2g' | '3g' | '4g'
+    readonly effective_type?: 'slow-2g' | '2g' | '3g' | '4g'
     /**
      * Cellular connectivity properties
      */