+++ packages/rum-core/test/mockContexts.ts
@@ -11,6 +11,8 @@ export function mockUrlContexts(fakeLocation: Location = location): UrlContexts 
       url: fakeLocation.href,
       referrer: document.referrer,
     }),
+    getAllEntries: () => [],
+    getDeletedEntries: () => [],
     stop: noop,
   }
 }
@@ -20,6 +22,7 @@ export function mockViewHistory(view?: Partial<ViewHistoryEntry>): ViewHistory {
     findView: () => view as ViewHistoryEntry,
     stop: noop,
     getAllEntries: () => [],
+    getDeletedEntries: () => [],
   }
 }